requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
alabaster==0.7.12
antlr4-python3-runtime==4.9.3
anyascii==0.3.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==20.2.0
Babel==2.10.3
backcall==0.2.0
blessings==1.7
blis==0.4.1
boto3==1.24.93
botocore==1.27.93
cachetools==5.2.0
catalogue==2.0.8
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
chatty-goose==0.2.0
click==8.1.3
coloredlogs==14.0
contractions==0.1.72
cymem==2.0.6
Cython==0.29.32
datasets==2.2.1
decorator==5.1.1
dill==0.3.5.1
docformatter==1.3
docutils==0.15.2
emoji==0.5.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.3
evaluate==0.3.0
executing==1.0.0
fairscale==0.4.12
faiss-cpu==1.7.2
faiss-gpu==1.7.2
filelock==3.8.0
flake8==3.7.8
flake8-bugbear==19.8.0
flatbuffers==2.0.7
frozenlist==1.3.1
fsspec==2022.2.0
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20220512
gast==0.4.0
gitdb==4.0.9
gitdb2==2.0.5
GitPython==3.0.3
google-api-core==2.10.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==1.43.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
gpustat==0.6.0
grpcio==1.48.1
h5py==3.7.0
huggingface-hub==0.10.1
humanfriendly==10.0
hydra-core==1.2.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.2.0
importlib-resources==5.10.0
iniconfig==1.1.1
iopath==0.1.10
ipython==7.31.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==0.14.1
jsonlines==1.2.0
keras==2.10.0
langcodes==3.3.0
libclang==14.0.6
lightgbm==3.3.2
llvmlite==0.39.0
Markdown==3.3.2
markdown-it-py==0.5.8
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mock==4.0.3
more-itertools==9.0.0
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
myst-parser==0.12.10
ninja==1.10.2.4
nltk==3.6.6
nmslib==2.1.1
numba==0.56.0
numpy==1.22.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.2.3
onnxruntime==1.12.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathy==0.6.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
pluggy==0.13.1
portalocker==2.6.0
preshed==3.0.7
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-gfm==1.0.2
py-rouge==1.1
pyahocorasick==1.4.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycodestyle==2.5.0
pydantic==1.8.2
pyflakes==2.1.1
pygaggle==0.0.3.1
Pygments==2.13.0
pyjnius==1.4.2
pyparsing==3.0.9
pyserini==0.18.0
pytest==5.3.2
pytest-datadir==1.3.1
pytest-regressions==2.1.1
python-dateutil==2.8.2
pytrec-eval==0.5
pytz==2022.2.1
PyYAML==5.4
pyzmq==18.1.0
regex==2022.8.17
requests==2.28.1
requests-mock==1.7.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==0.23.1
scipy==1.8.0
sentencepiece==0.1.97
sh==1.12.14
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
smmap2==3.0.1
snowballstemmer==2.2.0
spacy==3.3.0
spacy-legacy==3.0.10
spacy-loggers==1.0.3
Sphinx==2.2.2
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.5
stack-data==0.5.0
subword-nmt==0.3.7
sympy==1.11
tabulate==0.9.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
tensorflow==2.11.0rc0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
textsearch==0.0.24
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchtext==0.13.1
tornado==6.0.4
tqdm==4.62.3
traitlets==5.3.0
transformers==4.23.1
typer==0.4.2
typing-extensions==3.7.4.3
Unidecode==1.1.1
untokenize==0.1.1
urllib3==1.26.5
wasabi==0.10.1
wcwidth==0.2.5
websocket-client==0.56.0
websocket-server==0.6.4
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xxhash==3.0.0
yacs==0.1.8
yarl==1.8.1
zipp==3.8.1
